feat: Add hack for redirecting to correct page in NotFound component
This commit adds a hack in the NotFound component to redirect the user to the correct page when using a magic link. This hack is only necessary in production and not needed in development. It improves the user experience by ensuring that users are directed to the appropriate page when encountering a 404 error.